Dynamic Multiple Choice Question Form created.
Using react useState, I was able to create a dynamic form field. This allows the user to add fields for new questions as they would like while also retaining the user information in each field even if new fields are added. The option is limited to client request at a maximum of 20 questions and will be refactored to include TF questions in the next commit. The total will be calculated via both TF and MC # of fields, a pop up message will occur if the user attempts to go beyond the set maximum.