Create Quiz Functionality Complete.
I have first added form validations for both the MC and TF Questions. This ensures that the quiz fields will be filled out at the very least.

The Create Quiz button properly ensures both sections are validated as well as validating the quiz name.

Dynamic rendering of the MC and TF Fields is complete. Maximum questions added to be a combination of both # of MC Questions as well as TF Questions.

Database interaction will be complete by tomorrow.